requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==45.0.6
Django==4.2.23
django-allauth==65.11.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
